# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HSP Dispatch Service",
    "version": "0.1.0"
  },
  "paths": {
    "/api/dispatch/v1/workers/available": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "List available workers",
        "operationId": "list_available_workers_api_dispatch_v1_workers_available_get",
        "parameters": [
          {
            "name": "service_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Type"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "at_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At Time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAvailableWorkersResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business validation failed."
          },
          "503": {
            "description": "Dependent service unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dispatch/v1/dispatches/manual": {
      "post": {
        "tags": [
          "dispatch"
        ],
        "summary": "Manually assign order",
        "operationId": "manual_assign_api_dispatch_v1_dispatches_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business validation failed."
          },
          "404": {
            "description": "Resource not found."
          },
          "409": {
            "description": "Order already has pending dispatch or worker unavailable."
          },
          "503": {
            "description": "Dependent service unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dispatch/v1/workers/{worker_id}/pending-dispatches": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "List worker pending dispatches",
        "operationId": "list_pending_dispatches_api_dispatch_v1_workers__worker_id__pending_dispatches_get",
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Worker id.",
              "title": "Worker Id"
            },
            "description": "Worker id."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchRecordListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business validation failed."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dispatch/v1/workers/{worker_id}/dispatches/{dispatch_id}/response": {
      "post": {
        "tags": [
          "dispatch"
        ],
        "summary": "Confirm worker response",
        "operationId": "confirm_worker_response_api_dispatch_v1_workers__worker_id__dispatches__dispatch_id__response_post",
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Worker id.",
              "title": "Worker Id"
            },
            "description": "Worker id."
          },
          {
            "name": "dispatch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Dispatch id.",
              "title": "Dispatch Id"
            },
            "description": "Dispatch id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerResponseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business validation failed."
          },
          "404": {
            "description": "Dispatch not found."
          },
          "409": {
            "description": "Dispatch already responded."
          },
          "503": {
            "description": "Dependent service unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dispatch/v1/orders/{order_id}/dispatch-history": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "Get order dispatch history",
        "operationId": "get_order_history_api_dispatch_v1_orders__order_id__dispatch_history_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Order id.",
              "title": "Order Id"
            },
            "description": "Order id."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchRecordListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business validation failed."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Healthz Healthz Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AvailableWorkerResponse": {
        "properties": {
          "worker_id": {
            "type": "string",
            "title": "Worker Id",
            "description": "Worker unique id."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Worker display name."
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "description": "Worker skill tags."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Worker status from worker-schedule service."
          }
        },
        "type": "object",
        "required": [
          "worker_id",
          "name",
          "skills",
          "status"
        ],
        "title": "AvailableWorkerResponse"
      },
      "DispatchRecordListResponse": {
        "properties": {
          "dispatches": {
            "items": {
              "$ref": "#/components/schemas/DispatchRecordResponse"
            },
            "type": "array",
            "title": "Dispatches"
          }
        },
        "type": "object",
        "title": "DispatchRecordListResponse"
      },
      "DispatchRecordResponse": {
        "properties": {
          "dispatch_id": {
            "type": "string",
            "title": "Dispatch Id",
            "description": "Dispatch record id."
          },
          "order_id": {
            "type": "string",
            "title": "Order Id",
            "description": "Order id."
          },
          "attempt_no": {
            "type": "integer",
            "title": "Attempt No",
            "description": "Dispatch attempt sequence per order."
          },
          "worker_id": {
            "type": "string",
            "title": "Worker Id",
            "description": "Assigned worker id."
          },
          "operator_id": {
            "type": "string",
            "title": "Operator Id",
            "description": "Operator (customer service staff) id."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Dispatch status: PENDING/ACCEPTED/REJECTED."
          },
          "assigned_at": {
            "type": "string",
            "title": "Assigned At",
            "description": "Assignment time in ISO-8601 format."
          },
          "responded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responded At",
            "description": "Worker response time in ISO-8601 format."
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason",
            "description": "Worker reject reason if rejected."
          }
        },
        "type": "object",
        "required": [
          "dispatch_id",
          "order_id",
          "attempt_no",
          "worker_id",
          "operator_id",
          "status",
          "assigned_at"
        ],
        "title": "DispatchRecordResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListAvailableWorkersResponse": {
        "properties": {
          "workers": {
            "items": {
              "$ref": "#/components/schemas/AvailableWorkerResponse"
            },
            "type": "array",
            "title": "Workers"
          }
        },
        "type": "object",
        "title": "ListAvailableWorkersResponse"
      },
      "ManualAssignRequest": {
        "properties": {
          "order_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Order Id"
          },
          "worker_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Worker Id"
          },
          "operator_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Operator Id"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "worker_id",
          "operator_id"
        ],
        "title": "ManualAssignRequest",
        "example": {
          "operator_id": "csr-001",
          "order_id": "order-1001",
          "worker_id": "worker-001"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkerResponseRequest": {
        "properties": {
          "response": {
            "type": "string",
            "pattern": "^(ACCEPT|REJECT)$",
            "title": "Response"
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason"
          }
        },
        "type": "object",
        "required": [
          "response"
        ],
        "title": "WorkerResponseRequest",
        "example": {
          "reject_reason": "busy with previous order",
          "response": "REJECT"
        }
      }
    }
  }
}